refactor: Split off ServerInitializer.
