+++ config/presets/setup.json
@@ -4,17 +4,8 @@
     {
       "@id": "urn:solid-server:default",
       "@type": "Setup",
-      "Setup:_serverFactory": {
-        "@id": "urn:solid-server:default:ServerFactory"
-      },
       "Setup:_initializer": {
         "@id": "urn:solid-server:default:Initializer"
-      },
-      "Setup:_base": {
-        "@id": "urn:solid-server:default:variable:baseUrl"
-      },
-      "Setup:_port": {
-        "@id": "urn:solid-server:default:variable:port"
       }
     },
     {
@@ -37,6 +28,15 @@
           },
           "AclInitializer:_aclManager": {
             "@id": "urn:solid-server:default:AclManager"
+          }
+        },
+        {
+          "@type": "ServerInitializer",
+          "ServerInitializer:_serverFactory": {
+            "@id": "urn:solid-server:default:ServerFactory"
+          },
+          "ServerInitializer:_port": {
+            "@id": "urn:solid-server:default:variable:port"
           }
         }
       ]
