+++ src/index.ts
@@ -17,6 +17,7 @@ export * from './init/AclInitializer';
 export * from './init/CliRunner';
 export * from './init/Initializer';
 export * from './init/LoggerInitializer';
+export * from './init/ServerInitializer';
 export * from './init/Setup';
 
 // LDP/HTTP/Metadata
